refactor: split file for minimum usage.
