+++ api.ts
@@ -1,1 +1,4 @@
 export * from './apis/default-api'
+export * from './apis/default-api-factory'
+export * from './apis/funcionnal-programing-interface'
+export * from './apis/axios-parameter-creator'
