+++ apis/axios-parameter-creator.ts
@@ -1,0 +1,3043 @@
+import type { AxiosRequestConfig } from 'axios'
+
+import { Configuration } from '../configuration'
+import { type RequestArgs, RequiredError } from '../base'
+import type { ReadMethodQueryBody } from '../models'
+
+/**
+ * DefaultApi - axios parameter creator
+ * @export
+ */
+export const DefaultApiAxiosParamCreator = function (
+  configuration?: Configuration,
+) {
+  return {
+    /**
+     *
+     * @summary get address info
+     * @param {string} addressHash Address hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getAddress: async (
+      addressHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getAddress.',
+        )
+      }
+      const localVarPath = `/addresses/{address_hash}`.replace(
+        `{${'address_hash'}}`,
+        encodeURIComponent(String(addressHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get blocks validated by address
+     * @param {string} addressHash Address hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getAddressBlocksValidated: async (
+      addressHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getAddressBlocksValidated.',
+        )
+      }
+      const localVarPath = `/addresses/{address_hash}/blocks-validated`.replace(
+        `{${'address_hash'}}`,
+        encodeURIComponent(String(addressHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get address coin balance history
+     * @param {string} addressHash Address hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getAddressCoinBalanceHistory: async (
+      addressHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getAddressCoinBalanceHistory.',
+        )
+      }
+      const localVarPath =
+        `/addresses/{address_hash}/coin-balance-history`.replace(
+          `{${'address_hash'}}`,
+          encodeURIComponent(String(addressHash)),
+        )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get address coin balance history by day
+     * @param {string} addressHash Address hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getAddressCoinBalanceHistoryByDay: async (
+      addressHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getAddressCoinBalanceHistoryByDay.',
+        )
+      }
+      const localVarPath =
+        `/addresses/{address_hash}/coin-balance-history-by-day`.replace(
+          `{${'address_hash'}}`,
+          encodeURIComponent(String(addressHash)),
+        )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get address counters
+     * @param {string} addressHash Address hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getAddressCounters: async (
+      addressHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getAddressCounters.',
+        )
+      }
+      const localVarPath = `/addresses/{address_hash}/counters`.replace(
+        `{${'address_hash'}}`,
+        encodeURIComponent(String(addressHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get address internal txs
+     * @param {string} addressHash Address hash
+     * @param {string} [filter]
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getAddressInternalTxs: async (
+      addressHash: string,
+      filter?: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getAddressInternalTxs.',
+        )
+      }
+      const localVarPath =
+        `/addresses/{address_hash}/internal-transactions`.replace(
+          `{${'address_hash'}}`,
+          encodeURIComponent(String(addressHash)),
+        )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      if (filter !== undefined) {
+        localVarQueryParameter['filter'] = filter
+      }
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get address logs
+     * @param {string} addressHash Address hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getAddressLogs: async (
+      addressHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getAddressLogs.',
+        )
+      }
+      const localVarPath = `/addresses/{address_hash}/logs`.replace(
+        `{${'address_hash'}}`,
+        encodeURIComponent(String(addressHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get blocks validated by address
+     * @param {string} addressHash Address hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getAddressTokenBalances: async (
+      addressHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getAddressTokenBalances.',
+        )
+      }
+      const localVarPath = `/addresses/{address_hash}/token-balances`.replace(
+        `{${'address_hash'}}`,
+        encodeURIComponent(String(addressHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get address token transfers
+     * @param {string} addressHash Address hash
+     * @param {string} [type]
+     * @param {string} [filter]
+     * @param {string} [token]
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getAddressTokenTransfers: async (
+      addressHash: string,
+      type?: string,
+      filter?: string,
+      token?: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getAddressTokenTransfers.',
+        )
+      }
+      const localVarPath = `/addresses/{address_hash}/token-transfers`.replace(
+        `{${'address_hash'}}`,
+        encodeURIComponent(String(addressHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      if (type !== undefined) {
+        localVarQueryParameter['type'] = type
+      }
+
+      if (filter !== undefined) {
+        localVarQueryParameter['filter'] = filter
+      }
+
+      if (token !== undefined) {
+        localVarQueryParameter['token'] = token
+      }
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get blocks validated by address
+     * @param {string} addressHash Address hash
+     * @param {string} [type]
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getAddressTokens: async (
+      addressHash: string,
+      type?: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getAddressTokens.',
+        )
+      }
+      const localVarPath = `/addresses/{address_hash}/tokens`.replace(
+        `{${'address_hash'}}`,
+        encodeURIComponent(String(addressHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      if (type !== undefined) {
+        localVarQueryParameter['type'] = type
+      }
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get address txs
+     * @param {string} addressHash Address hash
+     * @param {string} [filter]
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getAddressTxs: async (
+      addressHash: string,
+      filter?: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getAddressTxs.',
+        )
+      }
+      const localVarPath = `/addresses/{address_hash}/transactions`.replace(
+        `{${'address_hash'}}`,
+        encodeURIComponent(String(addressHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      if (filter !== undefined) {
+        localVarQueryParameter['filter'] = filter
+      }
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get address withdrawals
+     * @param {string} addressHash Address hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getAddressWithdrawals: async (
+      addressHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getAddressWithdrawals.',
+        )
+      }
+      const localVarPath = `/addresses/{address_hash}/withdrawals`.replace(
+        `{${'address_hash'}}`,
+        encodeURIComponent(String(addressHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get native coin holders list
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getAddresses: async (
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      const localVarPath = `/addresses`
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get block info
+     * @param {string} blockNumberOrHash Block number or hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getBlock: async (
+      blockNumberOrHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'blockNumberOrHash' is not null or undefined
+      if (blockNumberOrHash === null || blockNumberOrHash === undefined) {
+        throw new RequiredError(
+          'blockNumberOrHash',
+          'Required parameter blockNumberOrHash was null or undefined when calling getBlock.',
+        )
+      }
+      const localVarPath = `/blocks/{block_number_or_hash}`.replace(
+        `{${'block_number_or_hash'}}`,
+        encodeURIComponent(String(blockNumberOrHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get block txs
+     * @param {string} blockNumberOrHash Block number or hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getBlockTxs: async (
+      blockNumberOrHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'blockNumberOrHash' is not null or undefined
+      if (blockNumberOrHash === null || blockNumberOrHash === undefined) {
+        throw new RequiredError(
+          'blockNumberOrHash',
+          'Required parameter blockNumberOrHash was null or undefined when calling getBlockTxs.',
+        )
+      }
+      const localVarPath =
+        `/blocks/{block_number_or_hash}/transactions`.replace(
+          `{${'block_number_or_hash'}}`,
+          encodeURIComponent(String(blockNumberOrHash)),
+        )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get block withdrawals
+     * @param {string} blockNumberOrHash Block number or hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getBlockWithdrawals: async (
+      blockNumberOrHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'blockNumberOrHash' is not null or undefined
+      if (blockNumberOrHash === null || blockNumberOrHash === undefined) {
+        throw new RequiredError(
+          'blockNumberOrHash',
+          'Required parameter blockNumberOrHash was null or undefined when calling getBlockWithdrawals.',
+        )
+      }
+      const localVarPath = `/blocks/{block_number_or_hash}/withdrawals`.replace(
+        `{${'block_number_or_hash'}}`,
+        encodeURIComponent(String(blockNumberOrHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get blocks
+     * @param {string} [type]
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getBlocks: async (
+      type?: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      const localVarPath = `/blocks`
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      if (type !== undefined) {
+        localVarQueryParameter['type'] = type
+      }
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get indexing status
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getIndexingStatus: async (
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      const localVarPath = `/main-page/indexing-status`
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get internal txs
+     * @param {string} transactionHash Transaction hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getInternalTxs: async (
+      transactionHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'transactionHash' is not null or undefined
+      if (transactionHash === null || transactionHash === undefined) {
+        throw new RequiredError(
+          'transactionHash',
+          'Required parameter transactionHash was null or undefined when calling getInternalTxs.',
+        )
+      }
+      const localVarPath =
+        `/transactions/{transaction_hash}/internal-transactions`.replace(
+          `{${'transaction_hash'}}`,
+          encodeURIComponent(String(transactionHash)),
+        )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get json rpc url
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getJsonRpcUrl: async (
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      const localVarPath = `/config/json-rpc-url`
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get logs
+     * @param {string} transactionHash Transaction hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getLogs: async (
+      transactionHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'transactionHash' is not null or undefined
+      if (transactionHash === null || transactionHash === undefined) {
+        throw new RequiredError(
+          'transactionHash',
+          'Required parameter transactionHash was null or undefined when calling getLogs.',
+        )
+      }
+      const localVarPath = `/transactions/{transaction_hash}/logs`.replace(
+        `{${'transaction_hash'}}`,
+        encodeURIComponent(String(transactionHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get main page blocks
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getMainPageTokens: async (
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      const localVarPath = `/main-page/blocks`
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get main page txs
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getMainPageTxs: async (
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      const localVarPath = `/main-page/transactions`
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get market chart
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getMarketChart: async (
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      const localVarPath = `/stats/charts/market`
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get NFT instance by id
+     * @param {string} addressHash Address hash
+     * @param {number} id integer id
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getNftInstance: async (
+      addressHash: string,
+      id: number,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getNftInstance.',
+        )
+      }
+      // verify required parameter 'id' is not null or undefined
+      if (id === null || id === undefined) {
+        throw new RequiredError(
+          'id',
+          'Required parameter id was null or undefined when calling getNftInstance.',
+        )
+      }
+      const localVarPath = `/tokens/{address_hash}/instances/{id}`
+        .replace(`{${'address_hash'}}`, encodeURIComponent(String(addressHash)))
+        .replace(`{${'id'}}`, encodeURIComponent(String(id)))
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get transfers of NFT instance
+     * @param {string} addressHash Address hash
+     * @param {number} id integer id
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getNftInstanceTransfers: async (
+      addressHash: string,
+      id: number,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getNftInstanceTransfers.',
+        )
+      }
+      // verify required parameter 'id' is not null or undefined
+      if (id === null || id === undefined) {
+        throw new RequiredError(
+          'id',
+          'Required parameter id was null or undefined when calling getNftInstanceTransfers.',
+        )
+      }
+      const localVarPath = `/tokens/{address_hash}/instances/{id}/transfers`
+        .replace(`{${'address_hash'}}`, encodeURIComponent(String(addressHash)))
+        .replace(`{${'id'}}`, encodeURIComponent(String(id)))
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get transfers count of NFT instance
+     * @param {string} addressHash Address hash
+     * @param {number} id integer id
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getNftInstanceTransfersCount: async (
+      addressHash: string,
+      id: number,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getNftInstanceTransfersCount.',
+        )
+      }
+      // verify required parameter 'id' is not null or undefined
+      if (id === null || id === undefined) {
+        throw new RequiredError(
+          'id',
+          'Required parameter id was null or undefined when calling getNftInstanceTransfersCount.',
+        )
+      }
+      const localVarPath =
+        `/tokens/{address_hash}/instances/{id}/transfers-count`
+          .replace(
+            `{${'address_hash'}}`,
+            encodeURIComponent(String(addressHash)),
+          )
+          .replace(`{${'id'}}`, encodeURIComponent(String(id)))
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get NFT instances
+     * @param {string} addressHash Address hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getNftInstances: async (
+      addressHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getNftInstances.',
+        )
+      }
+      const localVarPath = `/tokens/{address_hash}/instances`.replace(
+        `{${'address_hash'}}`,
+        encodeURIComponent(String(addressHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get raw trace
+     * @param {string} transactionHash Transaction hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getRawTrace: async (
+      transactionHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'transactionHash' is not null or undefined
+      if (transactionHash === null || transactionHash === undefined) {
+        throw new RequiredError(
+          'transactionHash',
+          'Required parameter transactionHash was null or undefined when calling getRawTrace.',
+        )
+      }
+      const localVarPath = `/transactions/{transaction_hash}/raw-trace`.replace(
+        `{${'transaction_hash'}}`,
+        encodeURIComponent(String(transactionHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get read methods
+     * @param {string} addressHash Address hash
+     * @param {string} [isCustomAbi]
+     * @param {string} [from]
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getReadMethods: async (
+      addressHash: string,
+      isCustomAbi?: string,
+      from?: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getReadMethods.',
+        )
+      }
+      const localVarPath =
+        `/smart-contracts/{address_hash}/methods-read`.replace(
+          `{${'address_hash'}}`,
+          encodeURIComponent(String(addressHash)),
+        )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      if (isCustomAbi !== undefined) {
+        localVarQueryParameter['is_custom_abi'] = isCustomAbi
+      }
+
+      if (from !== undefined) {
+        localVarQueryParameter['from'] = from
+      }
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get read methods proxy
+     * @param {string} addressHash Address hash
+     * @param {string} [isCustomAbi]
+     * @param {string} [from]
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getReadMethodsProxy: async (
+      addressHash: string,
+      isCustomAbi?: string,
+      from?: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getReadMethodsProxy.',
+        )
+      }
+      const localVarPath =
+        `/smart-contracts/{address_hash}/methods-read-proxy`.replace(
+          `{${'address_hash'}}`,
+          encodeURIComponent(String(addressHash)),
+        )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      if (isCustomAbi !== undefined) {
+        localVarQueryParameter['is_custom_abi'] = isCustomAbi
+      }
+
+      if (from !== undefined) {
+        localVarQueryParameter['from'] = from
+      }
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get smart contract
+     * @param {string} addressHash Address hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getSmartContract: async (
+      addressHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getSmartContract.',
+        )
+      }
+      const localVarPath = `/smart-contracts/{address_hash}`.replace(
+        `{${'address_hash'}}`,
+        encodeURIComponent(String(addressHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get verified smart contracts
+     * @param {string} [q]
+     * @param {string} [filter]
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getSmartContracts: async (
+      q?: string,
+      filter?: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      const localVarPath = `/smart-contracts`
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      if (q !== undefined) {
+        localVarQueryParameter['q'] = q
+      }
+
+      if (filter !== undefined) {
+        localVarQueryParameter['filter'] = filter
+      }
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get verified smart contracts counters
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getSmartContractsCounters: async (
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      const localVarPath = `/smart-contracts/counters`
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get state changes
+     * @param {string} transactionHash Transaction hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getStateChanges: async (
+      transactionHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'transactionHash' is not null or undefined
+      if (transactionHash === null || transactionHash === undefined) {
+        throw new RequiredError(
+          'transactionHash',
+          'Required parameter transactionHash was null or undefined when calling getStateChanges.',
+        )
+      }
+      const localVarPath =
+        `/transactions/{transaction_hash}/state-changes`.replace(
+          `{${'transaction_hash'}}`,
+          encodeURIComponent(String(transactionHash)),
+        )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get stats counters
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getStats: async (
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      const localVarPath = `/stats`
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get token info
+     * @param {string} addressHash Address hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getToken: async (
+      addressHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getToken.',
+        )
+      }
+      const localVarPath = `/tokens/{address_hash}`.replace(
+        `{${'address_hash'}}`,
+        encodeURIComponent(String(addressHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get token holders
+     * @param {string} addressHash Address hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getTokenCounters: async (
+      addressHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getTokenCounters.',
+        )
+      }
+      const localVarPath = `/tokens/{address_hash}/counters`.replace(
+        `{${'address_hash'}}`,
+        encodeURIComponent(String(addressHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get token holders
+     * @param {string} addressHash Address hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getTokenHolders: async (
+      addressHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getTokenHolders.',
+        )
+      }
+      const localVarPath = `/tokens/{address_hash}/holders`.replace(
+        `{${'address_hash'}}`,
+        encodeURIComponent(String(addressHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get token instance holders
+     * @param {string} addressHash Address hash
+     * @param {number} id integer id
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getTokenInstanceHolders: async (
+      addressHash: string,
+      id: number,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getTokenInstanceHolders.',
+        )
+      }
+      // verify required parameter 'id' is not null or undefined
+      if (id === null || id === undefined) {
+        throw new RequiredError(
+          'id',
+          'Required parameter id was null or undefined when calling getTokenInstanceHolders.',
+        )
+      }
+      const localVarPath = `/tokens/{address_hash}/instances/{id}/holders`
+        .replace(`{${'address_hash'}}`, encodeURIComponent(String(addressHash)))
+        .replace(`{${'id'}}`, encodeURIComponent(String(id)))
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get token token transfers
+     * @param {string} addressHash Address hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getTokenTokenTransfers: async (
+      addressHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getTokenTokenTransfers.',
+        )
+      }
+      const localVarPath = `/tokens/{address_hash}/transfers`.replace(
+        `{${'address_hash'}}`,
+        encodeURIComponent(String(addressHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get token transfers
+     * @param {string} transactionHash Transaction hash
+     * @param {string} [type]
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getTokenTransfers: async (
+      transactionHash: string,
+      type?: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'transactionHash' is not null or undefined
+      if (transactionHash === null || transactionHash === undefined) {
+        throw new RequiredError(
+          'transactionHash',
+          'Required parameter transactionHash was null or undefined when calling getTokenTransfers.',
+        )
+      }
+      const localVarPath =
+        `/transactions/{transaction_hash}/token-transfers`.replace(
+          `{${'transaction_hash'}}`,
+          encodeURIComponent(String(transactionHash)),
+        )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      if (type !== undefined) {
+        localVarQueryParameter['type'] = type
+      }
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get tokens list
+     * @param {string} [q]
+     * @param {string} [type]
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getTokensList: async (
+      q?: string,
+      type?: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      const localVarPath = `/tokens`
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      if (q !== undefined) {
+        localVarQueryParameter['q'] = q
+      }
+
+      if (type !== undefined) {
+        localVarQueryParameter['type'] = type
+      }
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get tx info
+     * @param {string} transactionHash Transaction hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getTx: async (
+      transactionHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'transactionHash' is not null or undefined
+      if (transactionHash === null || transactionHash === undefined) {
+        throw new RequiredError(
+          'transactionHash',
+          'Required parameter transactionHash was null or undefined when calling getTx.',
+        )
+      }
+      const localVarPath = `/transactions/{transaction_hash}`.replace(
+        `{${'transaction_hash'}}`,
+        encodeURIComponent(String(transactionHash)),
+      )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get txs
+     * @param {string} [filter]
+     * @param {string} [type]
+     * @param {string} [method]
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getTxs: async (
+      filter?: string,
+      type?: string,
+      method?: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      const localVarPath = `/transactions`
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      if (filter !== undefined) {
+        localVarQueryParameter['filter'] = filter
+      }
+
+      if (type !== undefined) {
+        localVarQueryParameter['type'] = type
+      }
+
+      if (method !== undefined) {
+        localVarQueryParameter['method'] = method
+      }
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get txs chart
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getTxsChart: async (
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      const localVarPath = `/stats/charts/transactions`
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get withdrawals
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getWithdrawals: async (
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      const localVarPath = `/withdrawals`
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get write methods
+     * @param {string} addressHash Address hash
+     * @param {string} [isCustomAbi]
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getWriteMethods: async (
+      addressHash: string,
+      isCustomAbi?: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getWriteMethods.',
+        )
+      }
+      const localVarPath =
+        `/smart-contracts/{address_hash}/methods-write`.replace(
+          `{${'address_hash'}}`,
+          encodeURIComponent(String(addressHash)),
+        )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      if (isCustomAbi !== undefined) {
+        localVarQueryParameter['is_custom_abi'] = isCustomAbi
+      }
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary get write methods proxy
+     * @param {string} addressHash Address hash
+     * @param {string} [isCustomAbi]
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    getWriteMethodsProxy: async (
+      addressHash: string,
+      isCustomAbi?: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling getWriteMethodsProxy.',
+        )
+      }
+      const localVarPath =
+        `/smart-contracts/{address_hash}/methods-write-proxy`.replace(
+          `{${'address_hash'}}`,
+          encodeURIComponent(String(addressHash)),
+        )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      if (isCustomAbi !== undefined) {
+        localVarQueryParameter['is_custom_abi'] = isCustomAbi
+      }
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary query read method
+     * @param {ReadMethodQueryBody} body
+     * @param {string} addressHash Address hash
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    queryReadMethod: async (
+      body: ReadMethodQueryBody,
+      addressHash: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      // verify required parameter 'body' is not null or undefined
+      if (body === null || body === undefined) {
+        throw new RequiredError(
+          'body',
+          'Required parameter body was null or undefined when calling queryReadMethod.',
+        )
+      }
+      // verify required parameter 'addressHash' is not null or undefined
+      if (addressHash === null || addressHash === undefined) {
+        throw new RequiredError(
+          'addressHash',
+          'Required parameter addressHash was null or undefined when calling queryReadMethod.',
+        )
+      }
+      const localVarPath =
+        `/smart-contracts/{address_hash}/query-read-method`.replace(
+          `{${'address_hash'}}`,
+          encodeURIComponent(String(addressHash)),
+        )
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'POST',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      localVarHeaderParameter['Content-Type'] = 'application/json'
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+      const needsSerialization =
+        typeof body !== 'string' ||
+        localVarRequestOptions.headers['Content-Type'] === 'application/json'
+      localVarRequestOptions.data = needsSerialization
+        ? JSON.stringify(body !== undefined ? body : {})
+        : body || ''
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary search
+     * @param {string} [q]
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    search: async (
+      q?: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      const localVarPath = `/search`
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      if (q !== undefined) {
+        localVarQueryParameter['q'] = q
+      }
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+    /**
+     *
+     * @summary search redirect
+     * @param {string} [q]
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     */
+    searchRedirect: async (
+      q?: string,
+      options: AxiosRequestConfig = {},
+    ): Promise<RequestArgs> => {
+      const localVarPath = `/search/check-redirect`
+      // use dummy base URL string because the URL constructor only accepts absolute URLs.
+      const localVarUrlObj = new URL(localVarPath, 'https://example.com')
+      let baseOptions
+      if (configuration) {
+        baseOptions = configuration.baseOptions
+      }
+      const localVarRequestOptions: AxiosRequestConfig = {
+        method: 'GET',
+        ...baseOptions,
+        ...options,
+      }
+      const localVarHeaderParameter = {} as any
+      const localVarQueryParameter = {} as any
+
+      if (q !== undefined) {
+        localVarQueryParameter['q'] = q
+      }
+
+      const query = new URLSearchParams(localVarUrlObj.search)
+      for (const key in localVarQueryParameter) {
+        query.set(key, localVarQueryParameter[key])
+      }
+      for (const key in options.params) {
+        query.set(key, options.params[key])
+      }
+      localVarUrlObj.search = new URLSearchParams(query).toString()
+      let headersFromBaseOptions =
+        baseOptions && baseOptions.headers ? baseOptions.headers : {}
+      localVarRequestOptions.headers = {
+        ...localVarHeaderParameter,
+        ...headersFromBaseOptions,
+        ...options.headers,
+      }
+
+      return {
+        url:
+          localVarUrlObj.pathname + localVarUrlObj.search + localVarUrlObj.hash,
+        options: localVarRequestOptions,
+      }
+    },
+  }
+}
